Avoid warnings from recent and upcoming Clang and GCC releases
Prevent GCC 10 strncpy() warning by telling the compiler that
cram_file_def::file_id is not a C-style NUL-terminated string.

GCC 10 warns for `int_var = uint32_var = (uint32_t) -1`.

Clang 11 warns that float doesn't have enough precision to represent
RAND_MAX (0x7fffffff) exactly. Recode using uint64_t arithmetic rather
than (32-bit) float arithmetic.

+++ test/test-regidx.c
@@ -189,9 +189,9 @@ void test_custom_payload(void)
 
 void get_random_region(uint32_t min, uint32_t max, uint32_t *beg, uint32_t *end)
 {
-    long int b = rand(), e = rand();
-    *beg = min + (float)b * (max-min) / RAND_MAX;
-    *end = *beg + (float)e * (max-*beg) / RAND_MAX;
+    uint64_t b = rand(), e = rand();
+    *beg = min + (b * (max-min)) / RAND_MAX;
+    *end = *beg + (e * (max-*beg)) / RAND_MAX;
 }
 
 void test_random(int nregs, uint32_t min, uint32_t max)
